significant refactor in preparation for macros
