+++ bespokeasm/assembler/model/__init__.py
@@ -7,12 +7,9 @@ import sys
 import yaml
 
 from bespokeasm import BESPOKEASM_VERSION_STR, BESPOKEASM_MIN_REQUIRED_STR
-from bespokeasm.assembler.line_identifier import LineIdentifier
-from bespokeasm.assembler.byte_code.assembled import AssembledInstruction
 from bespokeasm.assembler.keywords import ASSEMBLER_KEYWORD_SET
 from bespokeasm.assembler.model.instruction_set import InstructionSet
 from bespokeasm.assembler.model.operand_set import OperandSet, OperandSetCollection
-
 
 class AssemblerModel:
     _config: dict
@@ -129,6 +126,10 @@ class AssemblerModel:
     def instruction_mnemonics(self) -> list[str]:
         return list(self._instructions.keys())
 
+    @property
+    def instructions(self) -> InstructionSet:
+        return self._instructions
+
     def get_operand_set(self, operand_set_name: str) -> OperandSet:
         return self._operand_sets.get_operand_set(operand_set_name)
 
@@ -161,16 +162,3 @@ class AssemblerModel:
         for item in self.predefined_memory_blocks:
             results.append(item['name'])
         return results
-
-    def parse_instruction(self, line_id: LineIdentifier, instruction: str) -> AssembledInstruction:
-        instr_parts = instruction.strip().split(' ', 1)
-        mnemonic = instr_parts[0].lower()
-        if len(instr_parts) > 1:
-            operands = instr_parts[1]
-        else:
-            operands = ''
-
-        instr_obj = self._instructions.get(mnemonic)
-        if instr_obj is None:
-            sys.exit(f'ERROR: {line_id} - Unrecognized mnemonic "{mnemonic}"')
-        return instr_obj.generate_bytecode_parts(line_id, mnemonic, operands, self.endian, self.registers)
