+++ bespokeasm/assembler/model/instruction_base.py
@@ -1,8 +1,5 @@
 import sys
 
-from bespokeasm.assembler.line_identifier import LineIdentifier
-from bespokeasm.assembler.byte_code.assembled import AssembledInstruction
-
 class InstructionBase:
     def __init__(
                 self,
@@ -11,7 +8,8 @@ class InstructionBase:
                 registers: set[str]
             ) -> None:
         self._mnemonic = mnemonic
-
+        self._default_endian = default_endian
+        self._registers = registers
 
     def __repr__(self) -> str:
         return str(self)
@@ -22,17 +20,3 @@ class InstructionBase:
     @property
     def mnemonic(self) -> str:
         return self._mnemonic
-
-    def generate_bytecode_parts(
-        self,
-        line_id: LineIdentifier,
-        mnemonic: str,
-        operands: str,
-        default_endian: str,
-        register_labels: set[str],
-    ) -> AssembledInstruction:
-        if mnemonic != self.mnemonic:
-            # this shouldn't happen
-            sys.exit(f'ERROR: {line_id} - INTERNAL - Asked instruction {self} to parse mnemonic "{mnemonic}"')
-
-        sys.exit(f'INTERNAL ERROR: {line_id} - Instruction "{mnemonic}" is not implemented.')
