+++ bespokeasm/assembler/model/operand/types/enumeration_operand.py
@@ -87,5 +87,5 @@ class EnumerationOperand(OperandWithArgument):
                     )
             if bytecode_part is None and arg_part is None:
                 return None
-            return ParsedOperand(self, bytecode_part, arg_part)
+            return ParsedOperand(self, bytecode_part, arg_part, operand)
         return None
