+++ bespokeasm/utilities.py
@@ -27,3 +27,13 @@ def is_string_numeric(value_str: str) -> bool:
         return False
     match = re.match(PATTERN_NUMERIC_COMPILED, value_str.strip())
     return match is not None and len(match.groups()) > 0
+
+
+PATTERN_ALLOWED_LABELS = re.compile(
+        r'^(?!__|\.\.)(?:(?:\.|_|[a-zA-Z])[a-zA-Z0-9_]*)$',
+        flags=re.IGNORECASE|re.MULTILINE
+    )
+
+def is_valid_label(s: str):
+    res = re.search(PATTERN_ALLOWED_LABELS, s)
+    return (res is not None)
