+++ test/test_assembler_model.py
@@ -7,7 +7,7 @@ from bespokeasm.assembler.label_scope import LabelScope, LabelScopeType
 import bespokeasm.assembler.model.operand as A
 import bespokeasm.assembler.model.operand_set as AS
 from bespokeasm.assembler.model import AssemblerModel
-from bespokeasm.assembler.byte_code.assembled import AssembledInstruction
+from bespokeasm.assembler.model.instruction_parser import InstructioParser
 
 from test import config_files
 
@@ -47,87 +47,87 @@ class TestConfigObject(unittest.TestCase):
 
         test_line_id = LineIdentifier(1212, 'test_instruction_parsing')
 
-        pi1 = model1.parse_instruction(test_line_id, 'LDA $f')
+        pi1 = InstructioParser.parse_instruction(model1, test_line_id, 'LDA $f')
         self.assertEqual(pi1.byte_size, 1, 'assembled instruciton is 1 byte')
         self.assertEqual(pi1.get_bytes(TestConfigObject.label_values), bytearray([0x1F]), 'assembled instruction')
 
-        pi2 = model1.parse_instruction(test_line_id, 'add label1+5')
+        pi2 = InstructioParser.parse_instruction(model1, test_line_id, 'add label1+5')
         self.assertEqual(pi2.byte_size, 1, 'assembled instruciton is 1 byte')
         self.assertEqual(pi2.get_bytes(TestConfigObject.label_values), bytearray([0x27]), 'assembled instruction')
 
-        pi3 = model1.parse_instruction(test_line_id, 'out')
+        pi3 = InstructioParser.parse_instruction(model1, test_line_id, 'out')
         self.assertEqual(pi3.byte_size, 1, 'assembled instruciton is 1 byte')
         self.assertEqual(pi3.get_bytes(TestConfigObject.label_values), bytearray([0xE0]), 'assembled instruction')
 
         with pkg_resources.path(config_files, 'register_argument_exmaple_config.yaml') as fp:
             model2 = AssemblerModel(str(fp), 0)
 
-        piA = model2.parse_instruction(LineIdentifier(1,'test_mov_a_i'), 'mov a, i')
+        piA = InstructioParser.parse_instruction(model2, LineIdentifier(1,'test_mov_a_i'), 'mov a, i')
         self.assertEqual(piA.byte_size, 1, 'assembled instruciton is 1 byte')
         self.assertEqual(list(piA.get_bytes(TestConfigObject.label_values)), [0b01000010], 'assembled instruction')
 
-        piB = model2.parse_instruction(test_line_id, 'mov a,[$1120 + label1]')
+        piB = InstructioParser.parse_instruction(model2, test_line_id, 'mov a,[$1120 + label1]')
         self.assertEqual(piB.byte_size, 3, 'assembled instruciton is 3 byte')
         self.assertEqual(piB.get_bytes(TestConfigObject.label_values), bytearray([0b01000110, 0x22, 0x11]), 'assembled instruction')
 
-        piC = model2.parse_instruction(test_line_id, 'add i')
+        piC = InstructioParser.parse_instruction(model2, test_line_id, 'add i')
         self.assertEqual(piC.byte_size, 1, 'assembled instruciton is 1 byte')
         self.assertEqual(piC.get_bytes(TestConfigObject.label_values), bytearray([0b10111010]), 'assembled instruction')
 
-        piD = model2.parse_instruction(test_line_id, 'mov [$110D + (label1 + LABEL2)] , 0x88')
+        piD = InstructioParser.parse_instruction(model2, test_line_id, 'mov [$110D + (label1 + LABEL2)] , 0x88')
         self.assertEqual(piD.byte_size, 4, 'assembled instruciton is 4 byte')
         self.assertEqual(piD.get_bytes(TestConfigObject.label_values), bytearray([0b01110111,  0x88, 0xFF, 0x11]), 'arguments should be in reverse order')
 
-        piE = model2.parse_instruction(test_line_id, 'mov [sp - label1] , 0x88')
+        piE = InstructioParser.parse_instruction(model2, test_line_id, 'mov [sp - label1] , 0x88')
         self.assertEqual(piE.byte_size, 3, 'assembled instruciton is 3 byte')
         self.assertEqual(piE.get_bytes(TestConfigObject.label_values), bytearray([0b01101111, 0x88, 0b11111110]), 'arguments should be in reverse order')
 
-        piF = model2.parse_instruction(test_line_id, 'mov [sp+label1] , 0x88')
+        piF = InstructioParser.parse_instruction(model2, test_line_id, 'mov [sp+label1] , 0x88')
         self.assertEqual(piF.byte_size, 3, 'assembled instruciton is 3 byte')
         self.assertEqual(piF.get_bytes(TestConfigObject.label_values), bytearray([0b01101111, 0x88, 2]), 'arguments should be in reverse order')
 
-        piG = model2.parse_instruction(test_line_id, 'mov [sp] , 0x88')
+        piG = InstructioParser.parse_instruction(model2, test_line_id, 'mov [sp] , 0x88')
         self.assertEqual(piG.byte_size, 3, 'assembled instruciton is 3 byte')
         self.assertEqual(piG.get_bytes(TestConfigObject.label_values), bytearray([0b01101111, 0x88, 0]), 'arguments should be in reverse order')
 
-        piH = model2.parse_instruction(test_line_id, 'mov [$8000], [label1]')
+        piH = InstructioParser.parse_instruction(model2, test_line_id, 'mov [$8000], [label1]')
         self.assertEqual(piH.byte_size, 5, 'assembled instruciton is 5 byte')
         self.assertEqual(piH.get_bytes(TestConfigObject.label_values), bytearray([0b01110110, 2, 0, 0, 0x80]), 'arguments should be in reverse order')
 
-        piI = model2.parse_instruction(test_line_id, 'mov [mar], [label1]')
+        piI = InstructioParser.parse_instruction(model2, test_line_id, 'mov [mar], [label1]')
         self.assertEqual(piI.byte_size, 3, 'assembled instruciton is 3 byte')
         self.assertEqual(piI.get_bytes(TestConfigObject.label_values), bytearray([0b01100110, 2, 0]), 'no offset should be emitted for [mar]')
 
-        piJ = model2.parse_instruction(test_line_id, 'swap [$8000], [label1]')
+        piJ = InstructioParser.parse_instruction(model2, test_line_id, 'swap [$8000], [label1]')
         self.assertEqual(piJ.byte_size, 5, 'assembled instruciton is 3 byte')
         self.assertEqual(piJ.get_bytes(TestConfigObject.label_values), bytearray([0b11110110, 0, 0x80, 2, 0]), 'arguments should NOT be in reverse order')
 
-        piK = model2.parse_instruction(test_line_id, 'mov [sp+2], [sp+4]')
+        piK = InstructioParser.parse_instruction(model2, test_line_id, 'mov [sp+2], [sp+4]')
         self.assertEqual(piK.byte_size, 3, 'assembled instruciton is 3 byte')
         self.assertEqual(piK.get_bytes(TestConfigObject.label_values), bytearray([0b01101101, 4, 2]), 'arguments should be in reverse order')
 
-        piL = model2.parse_instruction(test_line_id, 'pop i')
+        piL = InstructioParser.parse_instruction(model2, test_line_id, 'pop i')
         self.assertEqual(piL.byte_size, 1, 'assembled instruciton is 1 byte')
         self.assertEqual(list(piL.get_bytes(TestConfigObject.label_values)), [0b00001010], 'pop to i')
 
-        piM = model2.parse_instruction(LineIdentifier(158,'test_pop_empty_arg'), 'pop')
+        piM = InstructioParser.parse_instruction(model2, LineIdentifier(158,'test_pop_empty_arg'), 'pop')
         self.assertEqual(piM.byte_size, 1, 'assembled instruciton is 1 byte')
         self.assertEqual(piM.get_bytes(TestConfigObject.label_values), bytearray([0b00001111]), 'just pop')
 
-        piN = model2.parse_instruction(test_line_id, 'mov a, [sp+2]')
+        piN = InstructioParser.parse_instruction(model2, test_line_id, 'mov a, [sp+2]')
         self.assertEqual(piN.byte_size, 2, 'assembled instruciton is 2 byte')
         self.assertEqual(piN.get_bytes(TestConfigObject.label_values), bytearray([0b01000101, 2]), 'just move [sp+2] into a')
 
         with self.assertRaises(SystemExit, msg='should error on unallowed operand combinations'):
-            model2.parse_instruction(test_line_id, 'mov a, a')
+            InstructioParser.parse_instruction(model2, test_line_id, 'mov a, a')
         with self.assertRaises(SystemExit, msg='[mar] should have no offset'):
-            model2.parse_instruction(test_line_id, 'mov [mar+2], [label1]')
+            InstructioParser.parse_instruction(model2, test_line_id, 'mov [mar+2], [label1]')
         with self.assertRaises(SystemExit, msg='should error due to too many operands'):
-            model2.parse_instruction(test_line_id, 'mov a, i, j')
+            InstructioParser.parse_instruction(model2, test_line_id, 'mov a, i, j')
         with self.assertRaises(SystemExit, msg='should error due to too many operands'):
-            model2.parse_instruction(test_line_id, 'nop 123')
+            InstructioParser.parse_instruction(model2, test_line_id, 'nop 123')
         with self.assertRaises(SystemExit, msg='should error due to too few operands'):
-            model2.parse_instruction(test_line_id, 'mov a')
+            InstructioParser.parse_instruction(model2, test_line_id, 'mov a')
 
     def test_bad_registers_in_configuratin(self):
         with pkg_resources.path(config_files, 'test_bad_registers_in_configuratin.yaml') as fp:
